frontend/pyproject.toml
[project]
name = "amber-home-frontend"
version = "0.1.0"
description = "Frontend application for Amber-Home energy monitoring"
# readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.28.0",
    "plotly>=5.15.0",
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.0",
    "pytz>=2023.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]